pyproject.toml
[tool.poetry]
name = "giftlink"
version = "0.1.0"
description = ""
authors = ["JhonataAugust0 <jhon.augustosilva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
fastapi = "^0.115.5"
pydantic = "^2.9.2"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.14.0"
python-dotenv = "^1.0.1"
mkdocs-material = "^9.5.44"
mkdocstrings = "^0.27.0"
mkdocstrings-python = "^1.12.2"
uvicorn = "^0.32.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
bcrypt = "^4.2.1"
emails = "^0.6"
aiosmtplib = "^3.0.2"
cryptography = "^44.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini.options]
pythonpath = "./app"
addopts = "--doctest-modules"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.blue]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 108

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve --dev-addr localhost:9999"

test = "pytest -s -x --cov=app/tests/ -vv"
post_test = "coverage html"